requirements.txt
requests>=2.32.0
python-dotenv>=1.1.0
pytest>=7.4.3
fastmcp>=2.10.5